backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
web3==6.11.3
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
oauth2client==4.1.3
pandas==2.1.3
python-multipart==0.0.6
aiofiles==23.2.1
asyncio==3.4.3

eth-account==0.9.0
eth-abi==4.2.1